Welcome
This application only runs as expected for Windows systems.  This is
because it uses the tracert command line utility on windows to map IP
addresses.

Each IP address on the top list, provides a list of all the possible
next hop IP addresses it can connect to on the lower list, if it states:
0 hidden hops.  If it does state a number of hidden hops, it means a
number of hops had not been resolved on the way to the stated IP
address.

The next step of the project would be to provide a visual stimuli of the
information gained by tracert. The visual surface has not been coded to
do anything yet.

The is a java app and requires the java launcher (on View02) , after
being compiled.